rewrote some setxyz code to aid the compiler.  Fixed bugs in sumdd, infile and cc_mpi regarding compiling for double precision.

+++ sumdd_m.f90
@@ -16,7 +16,7 @@ contains
 ! (no other optimisation options, not even -g)
 !  On APAC it works with default optimisation and on the NEC with -Cvopt.
 
-      integer, intent(in) :: len, itype
+      integer(kind=4), intent(in) :: len, itype
       complex, dimension(len), intent(in)  :: dra
       complex, dimension(len), intent(inout) :: drb
       real :: e, t1, t2
